requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
fastapi==0.135.1
fonttools==4.61.1
h11==0.16.0
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pillow==12.1.1
psycopg2-binary==2.9.11
pyarrow==23.0.1
pybaseball==2.2.7
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
PyJWT==2.11.0
PyNaCl==1.6.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
xgboost==3.2.0
